remove daoData context provider - moved any hooks that are used into global hooks folder - updated remaining imports (types and hooks)
