--- src/contexts/daoData/treasury/types.ts
@@ -1,41 +1,0 @@
-import { BigNumber } from 'ethers';
-
-export interface TokenEvent {
-  transactionHash: string;
-  blockNumber: number;
-}
-
-export interface TokenDepositEvent extends TokenEvent {
-  address: string;
-  amount: BigNumber;
-}
-
-export interface TokenWithdrawEvent extends TokenEvent {
-  addresses: string[];
-  amount: BigNumber;
-}
-export interface ERC721TokenEvent extends TokenEvent {
-  contractAddresses: string[];
-  tokenIds: BigNumber[];
-}
-export interface ERC20TokenEvent extends TokenEvent {
-  contractAddresses: string[];
-  amounts: BigNumber[];
-}
-
-export interface TreasuryAssetFungible {
-  name: string;
-  symbol: string;
-  decimals: number;
-  contractAddress: string;
-  totalAmount: BigNumber;
-  formatedTotal: string;
-}
-
-export interface TreasuryAssetNonFungible {
-  name: string;
-  symbol: string;
-  tokenId: BigNumber;
-  contractAddress: string;
-  tokenURI: string;
-}
