--- src/contexts/daoData/types/accessControlDAO.ts
@@ -1,7 +1,0 @@
-export type ACRoleListener = (
-  target: string,
-  functionDesc: string,
-  encodedSig: string,
-  role: string,
-  _: any
-) => void;
