--- src/contexts/daoData/types/daoLegacy.ts
@@ -1,7 +1,0 @@
-export type DAOCreationListener = (
-  _daoAddress: string,
-  _accessControl: string,
-  _sender: string,
-  _creator: string,
-  event: any
-) => Promise<void>;
