--- src/contexts/daoData/types/index.ts
@@ -1,3 +1,0 @@
-export * from './accessControlDAO';
-export * from './governorModule';
-export * from './token';
