+++ src/hooks/useSafeRace.ts
